m5stack: address final review feedback
- Change OV3660 CTRL_RUN and CTRL_STANDBY constants from i32 to u8
  to avoid silent truncation risk with 'as u8' casts
- Remove redundant normalized_mode variable in warn_if_camera_standby_settings_conflict
  to avoid double-normalization confusion

+++ devices/m5stack_unit_cam/src/hardware/camera/controller.rs
@@ -12,8 +12,8 @@ use super::ov3660_sequence::{
     REG_SYSTEM_CTRL0,
 };
 
-const OV3660_CTRL_RUN: u8 = super::ov3660_sequence::CTRL_RUN as u8;
-const OV3660_CTRL_STANDBY: u8 = super::ov3660_sequence::CTRL_STANDBY as u8;
+const OV3660_CTRL_RUN: u8 = super::ov3660_sequence::CTRL_RUN;
+const OV3660_CTRL_STANDBY: u8 = super::ov3660_sequence::CTRL_STANDBY;
 
 #[derive(Debug, Clone, Copy)] // Added Clone
 pub enum CustomFrameSize {
